+++ lib/flutter_ringtone_manager.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'flutter_ringtone_manager_platform_interface.dart';
+import 'ios_system_sounds.dart';
 
 class FlutterRingtoneManager {
   /// Plays the default ringtone of the device
@@ -22,5 +23,13 @@ class FlutterRingtoneManager {
   /// [assetPath] contains the [Uri] formatted [String] of the target asset
   FutureOr<void> playAudioAsset(String assetPath) {
     return FlutterRingtoneManagerPlatform.instance.playAudioAsset(assetPath);
+  }
+
+  /// Plays the default short sound which provides in iOS platform
+  /// based on the param value of [SystemSoundID]
+  /// Note: iOS SDK not provides those values as a predefined and it may change based new versions
+  FutureOr<void> playIosSystemSoundByID(SystemSoundID id) {
+    return FlutterRingtoneManagerPlatform.instance
+        .playIosSystemSoundByID(id.value);
   }
 }