+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+* Custom sound file can play which locates under assets folder in the root directory of the flutter app
+* API for Custom sound play method available for both iOS and Android platform.
+* All the default short sounds of iOS can play for iOS target apps.
+
 ## 0.1.1
 * Code format to meet dart standards
 * Dependencies upgrade to latest versions.